fix: updating google login
