+++ package.json
@@ -15,6 +15,7 @@
       "react-google-login": "^5.1.10",
       "react-hook-form": "^5.0.1",
       "react-loading-overlay": "^1.0.1",
+      "react-router": "^5.1.2",
       "react-router-dom": "^5.1.2",
       "react-scripts": "3.4.1",
       "react-spinners": "^0.8.1",
@@ -68,10 +69,11 @@
       "@types/query-string": "^6.3.0",
       "@types/react": "^16.9.23",
       "@types/react-dom": "^16.9.5",
-      "@types/react-router-dom": "^5.1.3",
+      "@types/react-native-loading-spinner-overlay": "^0.5.1",
+      "@types/react-router": "^5.1.7",
+      "@types/react-router-dom": "^5.1.5",
       "dotenv": "^8.2.0",
       "prettier": "1.19.1",
-      "@types/react-native-loading-spinner-overlay": "^0.5.1",
       "typescript": "^3.8.3"
    }
 }
